Moving the RunSettingsProviderExtensions to common. Also moving the updation of the runsettings to include extensions folder to ProxyDataCollectorManager
