+++ src/Microsoft.TestPlatform.Common/Utilities/RunSettingsProviderExtensions.cs
@@ -1,13 +1,11 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-namespace Microsoft.VisualStudio.TestPlatform.CommandLine.Processors.Utilities
+namespace Microsoft.VisualStudio.TestPlatform.Common.Utilities
 {
-    using System;
     using System.Diagnostics.CodeAnalysis;
     using System.IO;
     using System.Xml;
-    using System.Diagnostics.Contracts;
 
     using Microsoft.VisualStudio.TestPlatform.Utilities;
     using Microsoft.VisualStudio.TestPlatform.Common.Interfaces;
