+++ src/vstest.console/Processors/CollectArgumentProcessor.cs
@@ -8,9 +8,9 @@ namespace Microsoft.VisualStudio.TestPlatform.CommandLine.Processors
 
     using System.Globalization;
 
-    using Microsoft.VisualStudio.TestPlatform.CommandLine.Processors.Utilities;
     using Microsoft.VisualStudio.TestPlatform.Common;
     using Microsoft.VisualStudio.TestPlatform.Common.Interfaces;
+    using Microsoft.VisualStudio.TestPlatform.Common.Utilities;
     using Microsoft.VisualStudio.TestPlatform.ObjectModel;
     using Microsoft.VisualStudio.TestPlatform.ObjectModel.Utilities;
