+++ src/vstest.console/Processors/ResultsDirectoryArgumentProcessor.cs
@@ -8,9 +8,9 @@ namespace Microsoft.VisualStudio.TestPlatform.CommandLine.Processors
     using System.IO;
     using System.Security;
 
-    using Microsoft.VisualStudio.TestPlatform.CommandLine.Processors.Utilities;
     using Microsoft.VisualStudio.TestPlatform.Common;
     using Microsoft.VisualStudio.TestPlatform.Common.Interfaces;
+    using Microsoft.VisualStudio.TestPlatform.Common.Utilities;
     using CommandLineResources = Microsoft.VisualStudio.TestPlatform.CommandLine.Resources.Resources;
 
     /// <summary>
